requirements-dev.txt
pytest==6.2.5
GitPython==3.1.40
boto3==1.26.42
aws-xray-sdk==2.12.1
aws_lambda_powertools==2.6.0
tftest==1.8.5